Add --unsafe mode, add barebone safe mode to LLDB/GDB.
I had bigger ideas for this, but let's keep it simple for time constraints. I may change it in the future.

For now, a very restrictive safe mode: only a few no-side-effects functions allowed, plus print only allowing `[a-zA-Z0-9_ *]`.

Added the `--unsafe` flag for all 3 debuggers to disable any security.